+++ src/main/kotlin/com/github/copyinaction/auth/domain/RefreshToken.kt
@@ -1,6 +1,9 @@
 package com.github.copyinaction.auth.domain
 
 import com.github.copyinaction.common.domain.BaseEntity
+import com.github.copyinaction.common.exception.CustomException
+import com.github.copyinaction.common.exception.ErrorCode
+import com.github.copyinaction.config.jwt.JwtTokenProvider
 import jakarta.persistence.*
 import java.time.LocalDateTime
 
@@ -22,7 +25,23 @@ class RefreshToken(
     val expiryDate: LocalDateTime,
 ) : BaseEntity() {
 
+    companion object {
+        fun create(user: User, jwtTokenProvider: JwtTokenProvider): RefreshToken {
+            return RefreshToken(
+                user = user,
+                token = jwtTokenProvider.createRefreshToken(),
+                expiryDate = jwtTokenProvider.getRefreshTokenExpiryDate()
+            )
+        }
+    }
+
     fun isExpired(): Boolean {
         return LocalDateTime.now().isAfter(expiryDate)
+    }
+
+    fun validateNotExpired() {
+        if (isExpired()) {
+            throw CustomException(ErrorCode.EXPIRED_REFRESH_TOKEN)
+        }
     }
 }